+++ packages/chakra-ui/src/FieldTemplate/FieldTemplate.tsx
@@ -22,10 +22,9 @@ export default function FieldTemplate<
     displayLabel,
     hidden,
     label,
-    onRemovePropertyClick,
+    onKeyRename,
     onKeyRenameBlur,
-    onDropPropertyClick,
-    onKeyChange,
+    onRemoveProperty,
     readonly,
     registry,
     required,
@@ -55,10 +54,9 @@ export default function FieldTemplate<
       disabled={disabled}
       id={id}
       label={label}
-      onRemovePropertyClick={onRemovePropertyClick}
+      onKeyRename={onKeyRename}
       onKeyRenameBlur={onKeyRenameBlur}
-      onDropPropertyClick={onDropPropertyClick}
-      onKeyChange={onKeyChange}
+      onRemoveProperty={onRemoveProperty}
       readonly={readonly}
       required={required}
       schema={schema}