+++ packages/core/test/Form.test.jsx
@@ -143,8 +143,6 @@ describeRepeated('Form common', (createFormComponent) => {
           edit: true,
           uiSchema: {},
           fieldPathId: { $id: 'root', path: [] },
-          schemaValidationErrors: undefined,
-          schemaValidationErrorSchema: undefined,
           schemaUtils: sinon.match.object,
         });
       });
@@ -1974,8 +1972,6 @@ describeRepeated('Form common', (createFormComponent) => {
           fieldPathId: { $id: 'root', path: [] },
           schema: formProps.schema,
           uiSchema: {},
-          schemaValidationErrors: undefined,
-          schemaValidationErrorSchema: undefined,
           schemaUtils: sinon.match.object,
         });
       });