+++ packages/fluentui-rc/src/WrapIfAdditionalTemplate/WrapIfAdditionalTemplate.tsx
@@ -37,7 +37,7 @@ export default function WrapIfAdditionalTemplate<
     disabled,
     id,
     label,
-    onRemovePropertyClick,
+    onRemoveProperty,
     onKeyRenameBlur,
     readonly,
     required,
@@ -92,7 +92,7 @@ export default function WrapIfAdditionalTemplate<
           className='rjsf-object-property-remove'
           style={btnStyle}
           disabled={disabled || readonly}
-          onClick={onRemovePropertyClick}
+          onClick={onRemoveProperty}
           uiSchema={uiSchema}
           registry={registry}
         />