+++ packages/primereact/src/WrapIfAdditionalTemplate/WrapIfAdditionalTemplate.tsx
@@ -25,7 +25,7 @@ export default function WrapIfAdditionalTemplate<
   disabled,
   id,
   label,
-  onRemovePropertyClick,
+  onRemoveProperty,
   onKeyRenameBlur,
   readonly,
   required,
@@ -71,7 +71,7 @@ export default function WrapIfAdditionalTemplate<
           id={buttonId(id, 'remove')}
           className='rjsf-object-property-remove'
           disabled={disabled || readonly}
-          onClick={onRemovePropertyClick}
+          onClick={onRemoveProperty}
           registry={registry}
         />
       </div>