+++ packages/react-bootstrap/src/FieldTemplate/FieldTemplate.tsx
@@ -26,10 +26,9 @@ export default function FieldTemplate<
   disabled,
   label,
   hidden,
-  onRemovePropertyClick,
+  onKeyRename,
   onKeyRenameBlur,
-  onDropPropertyClick,
-  onKeyChange,
+  onRemoveProperty,
   readonly,
   required,
   schema,
@@ -52,10 +51,9 @@ export default function FieldTemplate<
       disabled={disabled}
       id={id}
       label={label}
-      onRemovePropertyClick={onRemovePropertyClick}
+      onKeyRename={onKeyRename}
       onKeyRenameBlur={onKeyRenameBlur}
-      onDropPropertyClick={onDropPropertyClick}
-      onKeyChange={onKeyChange}
+      onRemoveProperty={onRemoveProperty}
       readonly={readonly}
       required={required}
       schema={schema}