+++ packages/react-bootstrap/src/WrapIfAdditionalTemplate/WrapIfAdditionalTemplate.tsx
@@ -23,7 +23,7 @@ export default function WrapIfAdditionalTemplate<
   disabled,
   id,
   label,
-  onRemovePropertyClick,
+  onRemoveProperty,
   onKeyRenameBlur,
   readonly,
   required,
@@ -69,7 +69,7 @@ export default function WrapIfAdditionalTemplate<
           id={buttonId(id, 'remove')}
           className='rjsf-object-property-remove w-100'
           disabled={disabled || readonly}
-          onClick={onRemovePropertyClick}
+          onClick={onRemoveProperty}
           uiSchema={uiSchema}
           registry={registry}
         />